device type + code cleanup
