+++ src/main/java/com/nulp/mobilepartsshop/api/v1/ApiConstants.java
@@ -4,6 +4,8 @@ public class ApiConstants {
 
     public static final String GLOBAL_MAPPING_V1 = "/api/v1";
 
+    public static final String PART_MAPPING_V1 = GLOBAL_MAPPING_V1 + "/part";
+
     public static final String ADMIN_MAPPING = GLOBAL_MAPPING_V1 + "/admin";
 
     public static final String STAFF_MAPPING = GLOBAL_MAPPING_V1 + "/staff";
