--- src/main/java/com/nulp/mobilepartsshop/api/v1/manufacturer/dto/ManufacturerDto.java
@@ -1,19 +1,0 @@
-package com.nulp.mobilepartsshop.api.v1.manufacturer.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ManufacturerDto {
-
-    private Long id;
-
-    private String name;
-
-    private String logoUrl;
-}
