+++ src/main/java/com/nulp/mobilepartsshop/api/v1/part/dto/response/PartTypeResponse.java
@@ -1,0 +1,19 @@
+package com.nulp.mobilepartsshop.api.v1.part.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PartTypeResponse {
+
+    private Long id;
+
+    private String nameEn;
+
+    private String nameUk;
+}
