--- src/main/java/com/nulp/mobilepartsshop/core/entity/deviceType/DeviceType.java
@@ -1,27 +1,0 @@
-package com.nulp.mobilepartsshop.core.entity.deviceType;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "device_types")
-public class DeviceType {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    private String nameEn;
-
-    private String nameUk;
-}
