--- src/main/java/com/nulp/mobilepartsshop/core/entity/partType/PartType.java
@@ -1,27 +1,0 @@
-package com.nulp.mobilepartsshop.core.entity.partType;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "part_types")
-public class PartType {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    private String nameEn;
-
-    private String nameUk;
-}
