+++ src/main/java/com/nulp/mobilepartsshop/core/repository/part/DeviceTypeRepository.java
@@ -1,10 +1,10 @@
 package com.nulp.mobilepartsshop.core.repository.part;
 
-import com.nulp.mobilepartsshop.core.entity.part.PartType;
+import com.nulp.mobilepartsshop.core.entity.part.DeviceType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PartTypeRepository extends JpaRepository<PartType, Long> {
+public interface DeviceTypeRepository extends JpaRepository<DeviceType, Long> {
 
 }
